refactor(math): split TrajectoryMath.FrameTransform into its own file
